requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0  # For data processing in multi_agent_collaboration.py

# LangChain ecosystem
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.27
langchain-core>=0.1.0
langchain-experimental>=0.0.5
langgraph>=0.0.15
langsmith>=0.0.83  # Optional for tracking

# Tools and utilities
tavily-python>=0.2.8  # For search functionality
beautifulsoup4>=4.12.0  # For web scraping
ipython>=8.12.0  # For REPL functionality

# Type hinting and validation
typing-extensions>=4.8.0
pydantic>=2.5.2

# Path handling
pathlib>=1.0.1